Added and cleaned taxi trip data.
Selected an arbitrary day of taxi trips 2013-09-01 to run the min-cost flow
problem on and created a jupyter notebook to clean and discretize it. Generated
a CSV for both the cleaned and discrete inputs. Additionally, added a
draw_graph function to the code to visualize the flow graph on small example
instances.