Changed routing module and removed redundant modules.
Changed the init function of the routing module to no longer except trips and
arcs in the form of lists of tuples. Instead, there is a new function
create_dataframes which creates dataframes for these input types. The Data
Cleaning notebook was improved to include code for restriciting locations to
Manhattan and re-indexing of locations. A full run of that notebook generated
new CSVs for trips, nodes, and arcs. The addtional value feild was added to the
trips dataframe to give different values to each trip like revenue generated.
Lastly, the sub_problem and utility_function modules were removed.