Plot taxi routes and compute solution statistics.
Added the functionality to routing module to plot any given set of taxis path
on a map of Manhattan. Each taxi is color-coded and the opacity of the path
changes depending on if they are on a trip or not. Furthermore, added functions
to compute misc. statistics for each taxi and compute misc. aggregate
statistics as well.