Min-cost flow demo and changed draw_graph function.
Added a draft of a min-cost flow demo. Furthermore, changed the draw_graph
function in the routing module to have an optional parameter draw_all which
is true by default. If false, it only draws trip arcs and arcs with strictly
positive flow. This makes it easier to see taxi paths when there are lots of
edges in the graph.