Edited baseball elimination lab and max flow lab bug fix.
Created a max flow ford fulkerson implementation python module that can be used
in the baseball elimination lab. A bug in the update flow code was found and
fixed. The baseball elimination lab needs a few questions added to the end of
the lab to address a proof of elimination using the min-cut.