Improvements to the branch and bound lab.
The branch and bound (and knapsack) lab is not split in to the following parts:

(1) Branch and Bound: Some tricky questions including ones with a minmization
(2) Knapsack Problem: Using branch and bound model to solve knapsack
(3) Cutting planes: Identify cutting planes graphically and apply to knapsack.
(4) Geometry ot Branch and Bound: Will include 2d and 3d bnb visualizations

Might include a visualization of a knapsack specific branch and bound.