Overhauled Game Animation in penalty_kick_game
Added animation to the "automation" version of the penalty_kick_game. Animation in this track has been improved from the default track so that the goalie will be drawn at the position of the ball ONLY if the ball did not make it into the goal. Future improvements will be made to make the goalie only be drawn at the position of the ball if the ball is also aimed at the same location of the goalie, in order to further make clear the situation that took place during the game.


Former-commit-id: 50f6ee7d131392299ae29e266902be3ee3b10ab1