Improved Animation in Penalty Kick Game
Further improved animation for the "automation" track of the game so that the ball will visibly go out of the bounds of the goal if the striker missed and the goalie was not in the same location as the striker was aiming (E.G: The goalie stayed in the middle and the striker aimed at the right). If the striker does not score and the goalie stayed at the same location as the striker was aiming, then the goalie will be displayed at the same location as the ball to show that the ball was blocked.


Former-commit-id: ea206473851d94c38f9ae6f999ab07949e714fcf