Made suggestion to improve game_theory_lab_key flow
Identified a situation within the lab that tells a student to understand something before proceeding, then explains the exact same thing later on in the handout. Made a cell about the issue, suggesting that the explanation should be moved to earlier in the handout (before the student is asked to understand it before proceeding).


Former-commit-id: e9dedb3e91649abca3a1f12e7fb1249a2f6c6645