Added Suggestions to Improve game_theory_lab Consistency
Added two suggestion cells to game_theory_lab_key.ipynb. These suggestions are regarding the goalie's strategies for the game, and the solution to Q29/Q30.

Currently, the goalie for the game was defined to use fictitious play. As this means the goalie's strategy is supposed to counter the player's strategy, the goalie staying in the middle in response to the developed mixed strategy is correct, but possibly counterproductive to the lab.

Meanwhile, the answers to Q29/Q30 are technically incorrect, as the left footed striker instructions are not included, and therefore the instructions to the goalie are technically different.


Former-commit-id: c31573d836ab04b75a1d2a857a0227d74a6e9a42