Fixed constraints in lp modelling for full question without using dominance for game_theory_lab
Constraints seemed to use the wrong value for all left footed striker considerations. I fixed them to use the proper left footed striker values, and added comments for all constraints and variables. Unfortunately, this actually made the dual solution less clear in its results.

Currently, the dual solution still shows that the goalie will respond to the right footed striker in the same way, but the constraints that consider the goalie's actions for guarding the middle need to be added together to obtain the percentage. It appears that the left footed striker goalie instructions are still incorrect, due to the striker never being left footed.


Former-commit-id: 61999676b73c83c1756db7886f13438c9826e196