Fixed Minor Mistake in LP Comments for game_theory_lab
Fixed a minor typo in the LP Comments for the full LP solving the question without using dominance, where the comment had the incorrect chances for scoring written in (The code was still correct, so there have been no changes to the code). Additionally, a different version of the LP was tested, and although it obtained the same result and had less constraints, the dual was messier so I reverted back to this version.


Former-commit-id: 666c235b644eba990b2a90bcd98713da883fe06f