Preparation Changes For penalty_kick_game Computer Strategy Selection.
Preparation changes for the penalty_kick_game to support the player choosing the strategy for the computer to use for the goalie. While currently, the computer is only capable of using fictitious learning, these changes have added a dropdown menu that will be used for selecting the computer's strategy in the future. Changes were also made to the structure of the automation version's main game loop to make it easier to implement the mixed strategy decision making in the future.


Former-commit-id: f789b9f331c5b099122be54c2438c8c570aa47fa