Added First penalty_kick_game Game Data Display Figure
Added a figure used to display stats that gets updated throughout each iteration of the game. The figure added creates a custom bar chart showing the distribution of the player shot strategies, and the results obtained through each strategy. Each player option (E.g. Left-footed striker shooting left) has its own bar made up of four colors. The height of the bar corresponds to how many times the player chose that option, while the red portion of the bar corresponds to how many times the player scored while using that option, the blue portion of the bar corresponds to how many times the goalie blocked the ball by going left when the player chose that option, the green portion corresponds to how many times the goalie blocked the ball by going to the middle when the player chose that option, and the yellow portion corresponds to how many times the goalie blocked the ball by going to the right when the player chose that option.


Former-commit-id: 5af62f8c8aeb6525bc469bd046ac262dcc018518