Improved penalty_kick_game Game Data Display Figure 1
Added a legend to the first game data display figure, showing the meaning of each color in the figure.

Also added a custom HoverTool to the data display figure. This tool displays the data on a selected bar to the user, and should make reading the graph significantly easier.

Finally, improved how the actual figure works by adding in code to resize the y axis size according to the values contained within it.


Former-commit-id: 1368831aa49d8ffb6f1b1325acb82a6fc3eb2ebf