Formatting Changes And Code Improvement to penalty_kick_game
Made formatting changes and miscellaneous code changes to improve the readability of the code in penalty_kick_game.ipynb. This change was made before the inclusion of more game stat figures, as it should make it easier to work on the game stat figures, while also being a useful preparation change for any future collaborative work on the demo.


Former-commit-id: 9e70a33db1b03477e8a5e80c1cf137947d62b8a9