Added penalty_kick_game Game Stats Figure 2
Added a second game stats figure shown to the user at the end of the game. This figure plots the player's score against the iteration number of the game. It does not need a legend, but it has a custom hover tool, box and scroll zoom tools, and a pan tool.

The custom hover tool has been customized so that it clearly displays the stats, but also highlights the data point being hovered within the graph (The highlighted point is enlarged and its color is changed). On top of this, it uses a makeshift hitbox system (done using invisible glyphs ) to make it easier to select a data point in the plot to hover.

Finally, the plot auto adjusts it's height based off of the data points, so that it can make the most of the space allocated to it.


Former-commit-id: 02fcf1c20b27afac908fa3516f3137418e87fb0c