Added penalty_kick_game Game Stats Figure 3
Added the third game stats figure to the penalty kick game. This figure only displays if the selected strategy is set to "Fictitious_Play", and displays the perceived risks of each goalie position in a plot of the perceived risks over the iterations run. This plot also uses a custom HoverTool and makeshift hitbox system to more clearly highlight the data values to the user.


Former-commit-id: 7162e2dd965aa7a19203fb9fe7deb3e88b877566