penalty_kick_game Added Goalie Random Decision Strategy and Minor Bugfix
Added the Random strategy to the code for the main game loop. This makes it so that if the player selects the "Random" strategy in the menu dropdown, the goalie will select actions (Going left, middle, or right) at random (with an equal chance between options).

Also fixed a small bug that caused the decision tracking table to not display updates.


Former-commit-id: e7ff3edac8d61e12877cfce049ee1b7d63590a29