penalty_kick_game Created LP for Finding Optimal Mixed Counter Strategy
Created an LP that takes in 6 arguments (One for the chance associated with each of the possible striker choices. I.e. : Left footed, left side.) then outputs an array representing the optimal strategy for the goalie to use to minimize the chances of the striker scoring if they use the input mixed strategy. This function will hopefully be linked to a bokeh event so that it can be used to obtain the optimal mixed strategy for the goalie to take if the player selects the "Goalie_Cheats" option.


Former-commit-id: aec7d85d5d173b5cd7ffbc566180cac137481ebf