Removed LP for Finding Optimal Mixed Counter Strategy And b_make_counter Setup
Removed the LP made in the last commit, as bokeh does not support python callbacks being used in the way that the LP would require without a bokeh server. This should not cause many issues, because simplifying the LP reveals that it can be solved easily, so a CustomJS can be made to solve the specific problem instead. After removing the LP, preparation changes were made so that the b_make_counter button should only need the CustomJScode string to be added before it is functional.


Former-commit-id: 7c81a0a770de784256a7bcecbe8406abc125e3ca