penalty_kick_game Added Goalie_Cheats Goalie Strategy Into Main Game Loop
Added the goalie strategy corresponding to the Goalie_Cheats option in the strategy dropdown to the main game loop. The Goalie_Cheats strategy has the goalie only select the optimal action based off of the average risks of choosing each position. The average risks for this strategy are taken by having the computer directly take the chances of the striker choosing each of their actions, then calculating the chance that the striker scores for each pure strategy the goalie can take. This strategy is called "Goalie_Cheats" because the action of taking the striker's chances for each input is against the game's rules as stated in the lab handout.


Former-commit-id: 305b65b511a220d2f1901c8dc46db8fd8227959f