penalty_kick_game Moved Creation of grid1 to penalty_kick_automated_game
Moved the creation of the game's Gridplot grid1 to the new function format_layout contained within penalty_kick_automated_game.py. As this was the last game component creation section left in the penalty_kick_game Jupyter Notebook, it should now be possible to now move the entirety of the game's setup into penalty_kick_automated_game.py to be done through a single function call.


Former-commit-id: 24edd4b9f8c282acbdd1788ce7d91e5b6da9c601