penalty_kick_game Moved Game Setup and Creation into penalty_kick_automated_game
Moved the setup code and creation code for the game into the new function make_game contained within penalty_kick_automated_game.py. Also created new config classes for the main game figure, and each of the game stats figures. Each of the config classes stores the former parameters for the function used to setup the corresponding figure, and can be used to configure them in the same way as before.
If a config object is not provided to make_game, then make_game runs the setup for the corresponding figure using its default parameters.

Future updates will add in similar config classes for the remaining functions. These config classes were necessary in order to help organize the large amount of parameters the make_game function would otherwise have.


Former-commit-id: 2781939870aa19251ec5fc0f33ae3fe78794e0f7