penalty_kick_game Removed Repetitive Code In Stats Figure 3 hb_gc_code Strings
Created a function hb_gc_code that is used to obtain the v_func code string for the CustomJSTransforms used to get the y coordinates of the centers of the hitboxes. This function changes the line of code that differs for each set of hitboxes, eliminating the need for keeping the 6 very similar code strings that were previously being used.


Former-commit-id: f6d63a2fde9d6cb68f45fa6217c5eb19412e7606