penalty_kick_game Removed Repetitive Code In Stats Figure 3 ga Code Strings
created a function sfga_get_alpha_code that is used to obtain the v_func code string for the CustomJSTransforms used to get the alphas of the dots used by the figure to highlight the hovered plot points. This function changes the line of code that differs for each of the v_funcs used based off of the sfga parameter, eliminating the need for keeping the 6 very similar code strings that were previously being used.


Former-commit-id: 0ff2114b0f71fef49e9734269cb25febe82d7080