penalty_kick_game Stats Figures Polish Changes
Added buttons corresponding to the game stats figures, and increased the sizes of the figures. Only one figure is shown at a time now, but the buttons can be used to select which figure to display. Also increased the x axis range for stats figures 2 and 3 so that data points from the first and last iterations are more visible.


Former-commit-id: daf872f2a719f3349f852748ffd150dad7283c2d