penalty_kick_game Added Game Stats Figure 4
Added a 4th figure to the penalty_kick_game. This figure shows the score chance of each iteration in the game, so that the user can see how the score chances they had changed overtime. This can be useful not only in the lab, but also for testing, as it helps show whether or not certain goalie strategies are working as intended.


Former-commit-id: 8884f7c5681af337c3275e1dfca08853eed03a45