penalty_kick_game Removed Aim Sliders And Added Text Input Boxes
Replaced the aim sliders with text input boxes, so that the users could have better control of the chances they assigned to each possible striker foot, aim direction combination. This was actually a highly necessary change to make, as the sliders were not precise enough to allow the users to input a mixed strategy close enough to the solution obtained in the lab, causing some portions of the lab to act in unintended ways.

The text input system works as follows:

Once the Automate button has been clicked, the text inputs are shown.
All text inputs start with a default value of 0.
Three Divs are used to keep track of the status of setup steps necessary for starting the game. The three divs are as follows:
cpu_selected: default text value "0", used to track whether or not a strategy has been selected.
chances_valid: default text value "0", used to track whether the sum of the chances input by the user add up to 1, and all chances are between 0 to 1.
counter_made: default text value "1", used to track whether it is necessary to make a counter strategy before playing the game.

cpu_selected has its text value set to "1" when the user selects a strategy for the goalie.
chances_valid has its text value set to "0" whenever a change is made to the chances of each set of striker options happening. The input values are then checked to see if they pass the requirements (stated above), and if all of them are passed, chances_valid has its text set to "1".

counter_made has its text set to "0" whenever the goalie cheats strategy is selected. It has hits text set to "1" whenever a different strategy is selected, or the make_counter button is successfully completed.

Both the make_counter button and the start button are hidden unless their required divs are set to 1. The start button requires all three divs to have their texts set to 1, while the make_counter button requires the counter_made div to have its text value set to 0, and the chances_valid div to have its text value set to 1.

These requirements essentially enforce the following requirements on the visibility of each button:
start button: all chances are set correctly, strategy is selected, and if the strategy is goalie cheats, then make counter has been completed.
make_counter button: all chances are set correctly, strategy is selected and is goalie_cheats, make counter has not been completed.

All values input into the text inputs are also stored in the automation table so that they are used in the main game loop.

Due to an issue with javascript, the textinputs only check that the values provided are within a certain range of the value 1 (this range is very small, and is set through a variable at the top of the file). This is necessary because of how numbers work in javascript.


Former-commit-id: c0d1f00c83cff7f5c4786578941844eca69d16a6