Added game_theory_restructured_lab_key.ipynb
Created a new copy of the game_theory_lab_key containing a revised version of the lab that uses the new demo, and contains extra problems around other possible goalie strategies. The revised version has been kept separate in order to allow for the continued use of the base version as a tool for reference, and to allow for better testing of the lab before making any actual changes.


Former-commit-id: 8788bbbc4d19ae1e89136186e9acafbc5342671d