penalty_kick_game Added First Idea For Displaying Start Tips To User
Added 4 new divs that are displayed in the start button row if their conditions are met.
These divs are used to warn the user that something is preventing them from proceeding in the demo. The first div warns the user that they need to select a cpu strategy if they have not yet selected one from the dropdown. The second div warns the user that all chances must be set to a value less than 1. The third div warns the user in a similar fashion that all chance values must be set to a value greater than 0, and the final div warns the user that the chance values must add up to 1.

Each div is only shown to the user in the menu screen before starting the game. Each div is also only shown when the conditions they are warning against are in effect (E.g. the div warning against values greater than 1 is only shown if a value is greater than 1). The conditions in the menu are such that it should be impossible for a cautionary div and the start button to be displayed at the same time.


Former-commit-id: bbbd9507f9a724a3d9f8545ca87683d988043745