penalty_kick_game Added Automatic Iteration Advancement
Moved the bulk of the main game loop into a game iteration function. The game iteration function is never directly called. The game loop is now accessed through a new game loop async (asynchronous) function.

Also added a button to toggle automatically advancing between iterations, and a div to track whether or not the game is currently running an iteration.

The current game loop async function works first checks to make sure that no other game loop recursions are currently running (this prevents multiple iterations from running at the same time and causing errors). If no other recursions are running, then the game loop function updates the div to reflect that the game loop function is running, before running the game iteration. Following the completion of the game iteration, the game loop function checks to make sure that the iteration was not the last, and that the user has elected to automatically advance between iterations through the newly added button. If both conditions are fulfilled, then the game loop function will halt for a designated amount of time (currently set to 300 milliseconds, but the next update will be to make this amount adjustable), before recursing (It will update the div to show that no game loop recursions are running before immediately starting another game loop recursion that will run). Otherwise, the function ends after updating the div to show that no game loop recursions are running.

Changes were also made to the game components, so that the buttons used to run an iteration of the game (auto next and start automate buttons) are not activatable while a recursion is running (as an added insurance against the possibility of running multiple iterations at the same time).


Former-commit-id: 97e029838779990f2efd5fbdfaf75c14203eaefe