penalty_kick_game Main Game Creation Method Step Logging
Modified the game creation method penalty_kick_automated_game.main_game.MainGame.make_game()' to include a parameter 'log_steps' used to determine whether or not to print game creation progress during the construction of the game.

This change was made in advance of attempts to include parameters for configuring aspects of the game behaviour.


Former-commit-id: 748bfc71460d7adc6f2843b0690e1245f73b8ec0