penalty_kick_game Goalie Strategy Pre-designation
Modified the game creation method penalty_kick_automated_game.main_game.MainGame.make_game()' to include a parameter 'CPU_strategy' used to set the goalie strategy before the game start.

Setting the goalie strategy in such a way will remove the player's need to set the game strategy before starting the first iteration, and disable both the strategy dropdown (To prevent the player from changing the strategy), and the select CPU tip (As the player will no longer be required or able to select a strategy).


Former-commit-id: bbe86e2e24331b2938ace1f51a84df218a3eb297