penalty_kick_game Force Auto Advance Option
Modified the game creation method penalty_kick_automated_game.main_game.MainGame.make_game()' to include a parameter 'force_fast_forward' used to set whether or not the player should be forced to auto advance between iterations.

If allow_fast_forward is not set to false, setting 'forced_fast_forward' to True will set the value of the auto advance toggle button to true, and prevent it from being shown. This makes it so that the game will automatically advance iterations (As if the player clicked the toggle button), and also hide the toggle button so that the player cannot disable auto advancing.


Former-commit-id: d3ba7a68f55b9e591104bb86807682930306d6b4