penalty_kick_game Async Function Conversion For Neatness
Changed the game creation functions so that they happen asynchronously.

This allowed for the creation of the new function log_step(), a function used in a task created at the beginning of make_game(), that prints any strings added to the text_queue.

Moving the step logging to a separate async function allows for the game creation functions to simply put progress messages into the text queue to be processed by a step logging task (a task created to run log_step() while the game creation functions are running). This makes the code more readable by allowing for the removal of many parameter value checking 'if' statements.


Former-commit-id: 7f78c61f4d2ff68cf2aead024f5e0873bc0b3aca