Added .pre-commit-config.yaml for using nbstripout as a hook.
Added the necessary .pre-commit-config.yaml file for using nbstripout as a pre-commit hook to remove notebook metadata and outputs. nbstripout would still require the user to run a couple installation commands for the user to use it. Also added penalty_kick_game.ipynb as an example of a notebook being committed following the inclusion of nbstripout.


Former-commit-id: 2d4d930db6c36364322bbfeb0ce0b972f00e3884