Added clarity and some hints
Q1: Made it clearer that only the relationship between supply and demand is necessary not all of the sufficient conditions.
Part 3: Added a hint to look at the graph from before Q19 to aid understanding the formulation. The graph does not show node supply or edge capacity so it does not fully give away the answers Q17 or Q18. It does show where the edges from Q18 should go.
Q18: Added a hint that costs can be negative in the min-cost flow problem. For many other problems in the course, costs must be nonnegative, so the reminder can be helpful to avoid confusion.
Q23: Edited the template to make the arc syntax clearer. Specifically, made it clearer that it should be entered as a list of tuples.
Former-commit-id: 3d0f4121d3f223ffd8737c810e92743ff3f36422